track if bundles have actually changed, only notify when they have
In situations with many entrypoints in a single `esbuild.build`, esbuild
annoyingly rebuilds *everything* when any one file changes, even if only
one bundle depends on the change. So any `onEnd` plugins would get what
is effectively a bunch of noise of finished-but-unchanged outputs.

Because of that, I was doing one-build-per-entrypoint in my app. But I'm
finding that watch mode performance suffers badly with this approach.
CPU usage seems to scale roughly with number of builds, so once I ported
a big chunk of our endpoints watch mode CPU usage climbed to ~50% of a
CPU when just idling, which is unacceptable.

As a result I'm shifting my app back to many-entrypoints-per-build,
so now I need to work around the rebuild-everything behavior so this
plugin doesn't spew a bunch of pointless finished notifications. (This
still seems like less of a hassle than writing my own watch code.)

I considered reworking this plugin to require explicit calls of what
changed from an `onEnd` plugin within the actual app config, but I think
the overhead of calculating & tracking bundle hashes here is pretty
small. Keeping the plugin more plug-and-play is preferable as long as
performance is acceptable, so I'm going with this for now.