Fix client & server leaks
The listener was leaking server side (kept even after client conn was
closed), and the client (at least on some browsers) is not
garbage-collected after references are gone & must be manually
closed.

Without this, the client can cause a stampede of requests. A page
loads, tries to initialize one or a couple `EventSource` instances. They
fail. We create new ones to try in 10s, but the original ones still
exist and auto-retry seemingly in a second or two. So they fail again
and schedule *another* batch for 10s out. Wait 30s and you've got
thousands of `EventSource` instances competing.

Server side, there's a max number of listeners that stops things from
getting too out of hand, but this just meant I'd hit that limit way
earlier than expected with local dev tabs.